New shader compilation script, TV object
Rewrote the shader compilation script in python because it's a lot
simpler to understand and use than powershell for more complex things.

My plan for the future is to automate as much of the creation of
Vulkan-specific pipeline boilerplate as possible using a python script
that will generate the pipelines from shaders, this is a first step
in that direction.

Wanted to start working on lighting so I added an unbaked model, tv.obj,
for that purpose.