Python compilation script more generic
I migrated from Windows to Ubuntu recently and a lot of things need to
be changed to make the code build. I wanted to start with this
compilation script because it was a bit easier to work with and I wanted
to get used to working in this new environment.

Removed hard-coded paths and made it possible to specify input and
output directories.