ECS rework
I wanted my entity component system to be structured as vectors of
components for each statically declared archetype. This is why I started
this whole metaprogramming madness (which is a lot cooler than I
thought, I thought it was for nerds): I wanted to do as much work at
compile-time to have the generated code be as small and fast as possible
while still having the flexibility of runtime-created objects. The only
downsides are that all archetypes need to be declared before-hand at
compile-time and (at least as it stands now) you won't be able to remove
components from entities. Although both issues can be solved by thinking
ahead. I will have to implement a way to transition from one archetype
to another.