Better ArchetypeStorages, entity handles TODO
The ECS is taking shape (albeit all inside the main file; that will be
reworked next, this is just testing), but no system support and entity
handles are implemented wrong (reference invalidation galore). This is
good because it forces me to think about how the user will interact with
the ECS: DON'T use entity handles directly unless absolutely necessary -
for example, component hierarchies. You want to, ideally, only use
systems to modify component data to make best use of the cache. Handles
should only be created when you need them because they are not stored
as efficiently.