Added threadpool, basic job graph and scheduler
Users can schedule jobs and set checkpoints which will automatically
create tasks to run asynchronously in a threadpool. The features are
very rudimentary and I want to expand on some concepts like:

- not blocking the main thread while the scheduler runs
- make scheduling syntax more ergonomic/easy to read
- have a better job graph
- (much later) define the jobgraph through config files/GUI

The next step after all of this I think is making the rendering easier
to customize and expand. I want the program to compile the shaders and
generate code based on shader reflection